update tradeSteps to PlaceOrderSuccess & add warrings
